Moving delete-user tests to unit package
